+++ src/index.js
@@ -1,0 +1,13 @@
+
+import { OBJLoader2, OBJLoader2Parser } from "./loaders/OBJLoader2.js";
+import { OBJLoader2Parallel } from "./loaders/OBJLoader2Parallel.js";
+import { OBJ2LoaderWorker } from "./loaders/workerTaskManager/worker/tmOBJLoader2.js";
+import { MtlObjBridge } from "./loaders/utils/MtlObjBridge.js";
+
+export {
+	OBJLoader2,
+	OBJLoader2Parser,
+	OBJLoader2Parallel,
+	OBJ2LoaderWorker,
+	MtlObjBridge
+}