+++ src/loaders/workerTaskManager/worker/tmOBJLoader2.d.ts
@@ -1,0 +1,11 @@
+export class OBJ2LoaderWorker {
+    static buildStandardWorkerDependencies(threeJsLocation: any): ({
+        url: any;
+        code?: undefined;
+    } | {
+        code: any;
+        url?: undefined;
+    })[];
+    static init(context: any, id: any, config: any): void;
+    static execute(context: any, id: any, config: any): void;
+}