+++ src/index.js
@@ -1,0 +1,11 @@
+
+import { OBJLoader2, OBJLoader2Parser } from "./loaders/OBJLoader2.js";
+import { OBJLoader2Parallel } from "./loaders/OBJLoader2Parallel.js";
+import { MtlObjBridge } from "./loaders/utils/MtlObjBridge.js";
+
+export {
+	OBJLoader2,
+	OBJLoader2Parser,
+	OBJLoader2Parallel,
+	MtlObjBridge
+}