+++ src/loaders/utils/MtlObjBridge.d.ts
@@ -1,0 +1,4 @@
+export class MtlObjBridge {
+    static link(processResult: any, assetLoader: any): void;
+    static addMaterialsFromMtlLoader(materialCreator: any): {};
+}