+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 4.0.0-dev
 - Introduction of `WorkerTaskManager`
+- Updated Example `webgl_loader_obj2_options.html`: Expose parser options and logging configuration in UI
 
 ## 3.2.0
 - Changed structure of the repository to mimic three.js layout